requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.5
loguru>=0.7.2

# AI Services (optional - install if needed)
# anthropic>=0.7.8
# openai>=1.3.7
# elevenlabs>=0.2.0
